Revert "feat(history): prototype part1 (#1)"
This reverts commit 161b052016ab5837f42e2d2c8edfa6d0cd3fcf1b.
