+++ backend/backend/main.py
@@ -4,8 +4,6 @@ from fastapi.middleware.cors import CORSMiddleware
 from slowapi import Limiter, _rate_limit_exceeded_handler
 from slowapi.util import get_remote_address
 from slowapi.errors import RateLimitExceeded
-import uuid
-import datetime
 
 from .postgen import social_marketing
 
@@ -40,14 +38,14 @@ app.add_middleware(
 @limiter.limit(os.getenv("POSTS_CREATE_LIMIT") or "1/1minute")
 async def create_post(request: Request, data: mydomain.SocialMarketingPostRequest = Body(...)):
     text = await social_marketing.call_llm(data)
-    return {"id": uuid.uuid4(), "text": text, "createdAt": datetime.datetime.now()}
+    return text
 
 @app.post("/social-marketing/images")
 @limiter.limit(os.getenv("POSTS_CREATE_LIMIT") or "1/1minute")
 async def create_post(request: Request, data: mydomain.SocialMarketingImagetRequest = Body(...)):
-    resp = await social_marketing.call_llm_img(data)
-    return resp
+    text = await social_marketing.call_llm_img(data)
+    return text
 
 @app.get("/health")
 async def get_mquery(request: Request):
-    return {"status": "ok"}
+    return {"status": "ok"}
