+++ package.json
@@ -52,14 +52,11 @@
     "react-type-animation": "^3.2.0",
     "recharts": "^2.13.0",
     "remeda": "^2.15.2",
-    "rxdb": "15.35.0",
-    "rxjs": "7.8.0",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
     "url-join": "^5.0.0",
     "use-mutative": "^1.1.8",
-    "zod": "^3.23.8",
-    "zod-to-json-schema": "^3.23.3"
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.3.1",
