+++ src/api/index.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosResponse } from "axios"
+import axios from "axios"
 
 export const axiosMainInstance = axios.create({
   adapter: "fetch",
@@ -7,10 +7,7 @@ export const axiosMainInstance = axios.create({
 
 export async function createPost(body: any) {
   return axiosMainInstance
-    .post<
-      any,
-      AxiosResponse<{ id: string; text: string; createdAt: string }>
-    >("/social-marketing/posts", body)
+    .post("/social-marketing/posts", body)
     .then((res) => res.data)
 }
