+++ src/app/[lang]/social-marketing/page.tsx
@@ -1,7 +1,5 @@
-import { getEnvs } from "@/atoms/hooks"
-
-import SocialMarketingPage from "./SocialMarketingPage"
+import SocialMarketingPage from "./SocialMarketingPage";
 
 export default function SocialMarketingPageServer() {
-  return <SocialMarketingPage envs={getEnvs()} />
+  return <SocialMarketingPage />
 }
